+++ app/src/main/java/gaozhi/online/peoplety/service/record/GetFavoritesByUseridService.java
@@ -13,12 +13,12 @@ import gaozhi.online.peoplety.entity.Favorite;
 import gaozhi.online.peoplety.entity.Token;
 import gaozhi.online.peoplety.service.BaseApiRequest;
 import gaozhi.online.peoplety.service.NetConfig;
-import io.realm.Realm;
 
 /**
  * 获取收藏夹
  */
 public class GetFavoritesByUseridService extends BaseApiRequest<PageInfo<Favorite>> {
+
 
     public GetFavoritesByUseridService(OnDataListener<PageInfo<Favorite>> onDataListener) {
         super(NetConfig.favoriteBaseURL, Type.GET);
@@ -57,7 +57,8 @@ public class GetFavoritesByUseridService extends BaseApiRequest<PageInfo<Favorit
         //装入数据库
         getRealm().executeTransaction(realm -> {
             //删除过期缓存
-            realm.where(Favorite.class).lessThan("time", System.currentTimeMillis() - cathePeriod).findAll().deleteAllFromRealm();
+            if (realm.where(Favorite.class).findAll().size() > MIN_SIZE)
+                realm.where(Favorite.class).lessThan("time", System.currentTimeMillis() - cathePeriod).findAll().deleteAllFromRealm();
             List<Favorite> favorites = pageInfo.getList();
             favorites = realm.copyToRealmOrUpdate(favorites);
             pageInfo.setList(copyFromRealm(realm, favorites));
